Ensure simulator applies pity rules
